Add FileRepository service and RepositoryProtocol interface
Extract all file managing work into separate service called 'FileRepository' and replace private methods which strategy previously used with calls the instance methods of the repository. The strategy code, however, should interact with its interface, RepositoryProtocol.
Since the RepositoryProtocol rely on injected name of the file it will be operate on you should set up a property observer on 'fileNameURL' (aka 'fileURL'; more about that below) to refresh the file repository to reflect the actual state of the strategy.

Also refactor code of the strategies according to new usage of the 'fileURL' property. Now it's called 'fileNameURL' and should contain only file name and its extension. For example, URL(string: "MyLAMMPSFile.data") is the right usage.