Add copying data functionality
The 'copyDataToResults()' method should copy the given file (incapsulated inside of the RepositoryProtocol implementation) straight from the source folder into the results one. If copying can't be proceed because a file with the same name already exist in the results folder the method should replace the existing file with the new one.